# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "NodeJS Todo App Project",
    "contact": {
      "email": "NikenHananti.Puspasari@mitrais.com"
    }
  },
  "host": "localhost:3000",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/todo": {
      "get": {
        "tags": [
          "TodoController"
        ],
        "description": "Get all todo list",
        "operationId": "getTodo",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Todo"
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "TodoController"
        ],
        "description": "Create a new todo",
        "operationId": "postTodo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewTodo"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Todo"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Todo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "deadline": {
          "type": "string",
          "format": "date"
        },
        "done": {
          "type": "boolean"
        }
      }
    },
    "NewTodo": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "deadline": {
          "type": "string"
        },
        "done": {
          "type": "boolean"
        }
      }
    }
  }
}